renamed all survey_nn to survey_nn_v1
