+++ plaster/run/survey_nn_v1/survey_nn_v1_task.py
@@ -1,23 +1,23 @@
 from plaster.tools.pipeline.pipeline import PipelineTask
 from plaster.run.prep.prep_result import PrepResult
 from plaster.run.sim_v1.sim_v1_result import SimV1Result
-from plaster.run.survey_nn.survey_nn_params import SurveyNNParams
-from plaster.run.survey_nn.survey_nn_worker import survey_nn
+from plaster.run.survey_nn_v1.survey_nn_v1_params import SurveyNNV1Params
+from plaster.run.survey_nn_v1.survey_nn_v1_worker import survey_nn_v1
 
 
-class SurveyNNTask(PipelineTask):
+class SurveyNNV1Task(PipelineTask):
     def start(self):
-        survey_nn_params = SurveyNNParams(**self.config.parameters)
+        survey_nn_v1_params = SurveyNNV1Params(**self.config.parameters)
 
         prep_result = PrepResult.load_from_folder(self.inputs.prep)
         sim_result = SimV1Result.load_from_folder(self.inputs.sim_v2)
 
-        survey_nn_result = survey_nn(
-            survey_nn_params,
+        survey_nn_v1_result = survey_nn_v1(
+            survey_nn_v1_params,
             prep_result,
             sim_result,
             progress=self.progress,
             pipeline=self,
         )
 
-        survey_nn_result.save()
+        survey_nn_v1_result.save()
