+++ plaster/run/survey_nn_v1/zests/zest_survey_nn_v1.py
@@ -1,7 +1,7 @@
 from munch import Munch
 from plaster.run.error_model import ErrorModel
 from plaster.run.sim_v1.sim_v1_result import SimV1Result
-from plaster.run.survey_nn.survey_nn_worker import _euc_dist
+from plaster.run.survey_nn_v1.survey_nn_v1_worker import _euc_dist
 from plaster.tools.log.log import debug
 from zest import zest
